Non-production change to styleswitcher.js.
Was having a problem with the Codekit broswer refreshes and it wasn't
loading the stylesheet.  Problem is most likely a conflict between
setting the cookie for the stylesheet and the way Codekit refreshes the
page.  Disabled all cookie functions for now until the site is live to
try to prevent this issue in testing.