Every issue that I know of has been fixed.
Site has been tested and working 100% in the following browsers.
Chrome, Safari, Firefox, IE10.  Also all screen sizes are rendering the
page correctly.  IE7 displays the page exceptionally well.  It does not
render the border around the profile pic but the corners are rounded
and the drop shadow is present.  The gradients on the buttons don't
work, but I think they still look just fine.  Everything else displays
fine.
Next task is to clean up and organize CSS and HTML.  Create any needed
less variables.  Develop the rest of the pages and needed CSS using the
black style.  Once all pages are complete and site is fully working
with black style it can be cloned for the rest of the styles.